[L0] optimize # of event status queries through batching
Currently, for out-of-order immediate commandlists, each enqueued
operation creates an event on an internal vector to indicate
when that operation is completed. When that vector reaches
1k (by default) entries, each event is individually queried
for status and, if it completed, is cleaned up and removed.
This is causing latency spikes for enqueue operations and
reduces overall performance when using short-lived operations.

This patch introduces an event batching mechanism, where instead
of querying each event individually, the adapter will group
events in batches of 512 entries (by default) that will be
queried collectively with the use of a barrier.

Most of the credit goes to Michal Mrozek <michal.mrozek.intel.com>
for profiling the problem and proposing a solution.

This new mechanism is disabled by default. To use it, this patch
is adding three new env variables:

- UR_L0_IMMEDIATE_COMMANDLISTS_BATCH_EVENT_COMPLETIONS. Enables
  event completion batching. Defaults to off.
- UR_L0_IMMEDIATE_COMMANDLISTS_BATCH_MAX. Controls the maximum
  number of batches that will be created to store events. If
  there are no batches available, the adapter will synchronously
  wait for one to complete. Defaults to 10.
- UR_L0_IMMEDIATE_COMMANDLISTS_EVENTS_PER_BATCH. The number of
  events in each batch. Once a batch reaches this threshold,
  the adapter appends a barrier to the command list with the
  relevant batched events, and then another batch is selected
  to store incoming events. Defaults to 512.

Early results show appreciable reduction in enqueue tail latency,
but more experiments are needed.