add compute benchmark workflow for L0
This patch adds a script for running compute-benchmarks,
https://github.com/intel/compute-benchmarks/, and a corresponding
GH Actions workflow that runs those benchmarks when prompted
to do so with a comment, like so:

/benchmarks-level-zero --env UR_L0_IMMEDIATE_COMMANDLISTS_BATCH_EVENT_COMPLETIONS=1

Additional arguments can be appended to the end of the line. After the build if finished,
the results will be presented through a comment.

By default, benchmark results are ephemeral and discarded immediately. To save the
results of some run, use `--save <name>` to save the benchmark result data. To later
use it for comparison, use `--compare <name>`. By default, benchmarks are compared
against `baseline` results.

For now, this runs only a single scenario, api_overhead_benchmark_sycl with SubmitKernel test,
but will expand over time to cover more.