refactor loader lifetime management
This patch implements an atomic singleton class for managing
the lifecycle of the context objects inside of the loader.
This class ensures that the contexts always exist and lets the
loader manually destroy them on user request (during teardown).

Thanks to this change, the loader no longer relies on the order
of library constructors and destructors. It also gets us 90%
towards allowing the loader to be statically linked with the
application.