[L0 v2] Make L0 v2 implementation a seperate adapter
Initially, L0 v2 adapter was supposed to reside in a separate
namespace but be a part of legacy L0 adapter (with runtime option
to switch between executing on legacy or v2). However, this
turns out to require a lot of changes in the legacy code to
allow for function dispatching to legacy/v2 implementations of
queue, event, etc.

This approach allows us to keep the implementations separate while
still resuing files when appropriate (e.g. for adapter.cpp or
platform.cpp).